apps/server/requirements.txt
setuptools>=69.0.0
wheel>=0.42.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask-cors>=4.0.0
numpy>=2.1.0
scikit-learn>=1.4.0
python-docx>=1.0.0
tiktoken>=0.5.0